release: v1.0

+++ src/main/java/ta/simonitadepan/monitadepan/controller/BalitaController.java
@@ -57,9 +57,16 @@ BalitaController {
             @PathVariable Long id_balita,
             RedirectAttributes redirectAttributes
     ){
+        UserModel user = userService.getUserByUsername(SecurityContextHolder.getContext().getAuthentication().getName());
         BalitaModel balita = balitaService.getBalitaById(id_balita);
-        balitaService.deleteBalita(balita);
-        redirectAttributes.addFlashAttribute("msgDelete", "Data Balita Berhasil Dihapus!");
+        if (balita.getId_pengguna().equals(user)){
+            balitaService.deleteBalita(balita);
+            redirectAttributes.addFlashAttribute("msgDelete", "Data Balita Berhasil Dihapus!");
+        }
+        else{
+            redirectAttributes.addFlashAttribute("msgDelete", "Telah terjadi kesalahan!");
+        }
+
         return "redirect:/balita";
     }
 
@@ -86,7 +93,13 @@ BalitaController {
             RedirectAttributes redirectAttributes
     ){
         BalitaModel balita = balitaService.getBalitaById(id_balita);
-        balitaService.changeStatusBalita(balita);
+        UserModel user = userService.getUserByUsername(SecurityContextHolder.getContext().getAuthentication().getName());
+        System.out.println(balita.getId_pengguna().getUsername());
+        System.out.println(user.getUsername());
+        if (balita.getId_pengguna().getUsername() != user.getUsername()){
+            return "redirect:/";
+        }
+        balitaService.changeStatusBalita(balita, user);
         return "redirect:/";
     }
 }
